Update table content when data changes via  provider
fixes ismailyegnr/expandable_datatable#4

The problem was that the table content was created in the initState method instead of build method. Init state is only called once when constructing the widget while build is repeadedly called when things change in the UI allowing the view to update.
I also had to ajust the sorting mechanism a little so the sort state does not get lost during rebuilds